+++ lib/workato/connector/sdk/lookup_tables.rb
@@ -4,8 +4,6 @@
 require 'csv'
 require 'singleton'
 
-using Workato::Extension::HashWithIndifferentAccess
-
 module Workato
   module Connector
     module Sdk
@@ -42,7 +40,7 @@ module Workato
           @table_by_id ||= {}
           @table_by_name ||= {}
           data.each do |name, table|
-            table = HashWithIndifferentAccess.wrap(table)
+            table = Utilities::HashWithIndifferentAccess.wrap(table)
             rows = table['rows'].freeze
             @table_by_id[table['id'].to_i] = rows
             @table_by_name[name] = rows