+++ lib/workato/connector/sdk/settings.rb
@@ -3,13 +3,13 @@
 
 require 'active_support/encrypted_configuration'
 
-using Workato::Extension::HashWithIndifferentAccess
-
 module Workato
   module Connector
     module Sdk
       module SorbetTypes
-        SettingsHash = T.type_alias { T.any(HashWithIndifferentAccess, T::Hash[T.any(Symbol, String), T.untyped]) }
+        SettingsHash = T.type_alias do
+          T.any(ActiveSupport::HashWithIndifferentAccess, T::Hash[T.any(Symbol, String), T.untyped])
+        end
       end
 
       class Settings
@@ -100,13 +100,13 @@ module Workato
         end
 
         def read_encrypted_file
-          all_settings = HashWithIndifferentAccess.wrap(encrypted_configuration.config)
+          all_settings = Utilities::HashWithIndifferentAccess.wrap(encrypted_configuration.config)
 
           (name ? all_settings.fetch(name) : all_settings) || {}
         end
 
         def update_encrypted_file(new_settings)
-          all_settings = HashWithIndifferentAccess.wrap(encrypted_configuration.config)
+          all_settings = Utilities::HashWithIndifferentAccess.wrap(encrypted_configuration.config)
 
           merge_settings(all_settings, new_settings)