+++ lib/workato/connector/sdk/streams.rb
@@ -1,8 +1,6 @@
 # typed: strict
 # frozen_string_literal: true
 
-using Workato::Extension::HashWithIndifferentAccess
-
 module Workato
   module Connector
     module Sdk